Added last 10 user logins to Django admin
Last 10 user logins (along with their login times) now displayed on the
Django admin homepage.
