+++ agentex/datareduc.py
@@ -1,3 +1,6 @@
+'''
+All non render/page based views are stored here, rather than views.py.
+'''
 import json
 from django.utils.encoding import smart_unicode
 from django.core.serializers import serialize
@@ -170,3 +173,86 @@ def photometry(code,person,progress=False,admin=False):
     if admin:
         return normcals,stamps,indexes,cats
     return cals,normcals,list(sc),list(bg),dates,stamps,indexes,cats
+
+
+def measure_offset(d,person,basiccoord):
+    # Find the likely offset of this new calibrator compared to the basic ones and find any sources within 5 pixel radius search
+    finderid = d[0].event.finder
+    finderdp = Datapoint.objects.values_list('xpos','ypos').filter(user=person,data__id=finderid,pointtype='C',coorder__calid__lt=3).order_by('coorder__calid')
+    finder = basiccoord - array(finderdp)
+    t = transpose(finder)
+    xmean = mean(t[0])
+    ymean = mean(t[1])
+    return xmean,ymean
+
+def updatedisplay(request,code):
+    # Wipe all the validations for user and event
+    o = personcheck(request)
+    dc = DataCollection.objects.filter(person=o.user,planet=Event.objects.get(name=code),complete=True)
+    dc.update(display = False)
+    empty = True
+    formdata = request.POST
+    for i,val in formdata.items():
+        if i[4:] == val:
+            # Add validations back one by one
+            col = dc.filter(calid=val)
+            col.update(display= True)
+            empty = False
+    return empty
+
+def addvalidset(request,code):
+    o = personcheck(request)
+    calid = request.POST.get('calid','')
+    choice1 = request.POST.get('choice1','')
+    choice2 = request.POST.get('choice2','')
+    point = DataCollection.objects.filter(person=o.user,calid=calid,planet__name=code)
+    planet = Event.objects.filter(name=code)[0]
+    if choice1 and point and calid:
+        value = decisions[choice1]
+        source = point[0].source
+        old = Decision.objects.filter(person=o.user,planet=planet,source=source)
+        old.delete()
+        decision1 = Decision(source=source,
+                        value=value,
+                        person=o.user,
+                        planet=planet)
+        
+        if choice2:
+            value2 = decisions[choice2]
+            decision2 = Decision(source=source,
+                            value=value2,
+                            person=o.user,
+                            planet=planet,
+                            current=True)
+            decision2.save()
+        else:
+            decision1.current = True
+        decision1.save()
+        return False
+    else:
+        return True
+        
+@login_required
+def my_data(o,code):
+    data = []
+    sources = DataSource.objects.filter(event__name=code).order_by('timestamp')
+    points  = Datapoint.objects.filter(data__event__name=code,user=o.user)
+    for s in sources:
+        ps = points.filter(data=s)
+        myp = ps.filter(pointtype='S')
+        try:
+            mypoint = '%f' % myp[0].value
+        except:
+            mypoint = 'null'
+        cals = ps.filter(pointtype='C').values_list('value',flat=True).order_by('coorder')
+        line = {
+                'id'        : "%i" % s.id,
+                'date'      : s.timestamp.isoformat(" "),
+                'datestamp' : timegm(s.timestamp.timetuple())+1e-6*s.timestamp.microsecond,
+                'data'      : { 'source' : list(ps.filter(pointtype='S').values_list('value',flat=True)),
+                                'background' :  list(ps.filter(pointtype='B').values_list('value',flat=True)),
+                                'calibrator' :  list(cals),
+                            },
+                }
+        data.append(line)
+    return data,points
