+++ agentex/templates/agentex/agentex_info.html
@@ -1,6 +1,5 @@
-{% load admintags %}
 <div class="module">
-      <h2>Agent Exoplanet stats</h2>
+      <h2>Agent Exoplanet Stats</h2>
 <table class="bluetable">
    <tbody>
        <thead>
