+++ agentex/templates/agentex/get_last_login.html
@@ -1,0 +1,21 @@
+<div class="module">
+      <h2></br>Agent Exoplanet Users Stats</h2>
+<table class="bluetable">
+   <tbody>
+      <thead>
+        <tr>
+          <th>Parameter</th><th>Value</th>
+        </tr>
+      </thead>
+      <tr>
+        <td>Number of user logins in last week:</td><td>{{ num_users }}</td>
+      </tr>
+      <tr><th colspan=2>Last User Logins</th></tr>
+      {% for u in users %}
+      <tr>
+          <td>{{u.user}}</td><td>{{u.lastlogin}}</td>
+      </tr>
+      {% endfor %}
+   </tbody>
+</table>
+</div>
