+++ templates/admin/index.html
@@ -1,9 +1,9 @@
-{% extends "admin/base_site_old.html" %}
+{% extends "admin/base_site.html" %}
 {% load i18n %}
 {% load admintags %}
 {% load staticfiles %}
 
-{% block extrastyle %}{{ block.super }}<link rel="stylesheet" type="text/css" href="{% static 'css/dashboard.css' %}" />{% endblock %}
+{% block extrastyle %}{{ block.super }}<link rel="stylesheet" type="text/css" href="{% static "css2/dashboard.css" %}" />{% endblock %}
 
 {% block coltype %}colMS{% endblock %}
 
@@ -14,6 +14,7 @@
 {% block content %}
 <div id="content-main">
 {% agentex_info %}
+{% get_last_login %}
 {% if app_list %}
     {% for app in app_list %}
         <div class="module">
